requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-multiselectfield==0.1.12
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
